backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
python-ulid==2.7.0
python-dotenv==1.0.1
python-dateutil==2.9.0
reportlab==4.2.2
email-validator==2.2.0

# Testing
pytest==8.3.2
httpx==0.27.0
pytest-cov==5.0.0
